Add an init_callback parameter to the async pool
The purpose of this is for the supervisor to be able to dynamically
generate data that will be passed to the workers, an example being, the
supervisor might create ets tables. We keep this in the context of the
supervisor and its workers, to keep the abstraction isolated to this
supervision tree: an ets table may not make any sense outside of this
supervision tree, we might want the ets table not to even be named, so
that nobody outside of the tree will have access to it, only the workers
will get an opaque reference. We will also want the ets tables to be
cleaned when the supervision tree dies, as the data might not make sense
on a newly restarted tree anymore, or we will want the tree to die fast
if the ets table is lost. This init callback can also set global
configuration flags, or send messages, or log something, or define new
metrics, or many other things that for example the current MongooseIM
init callbacks for `mongoose_wpool` are already doing.