Use stored instead of generated pool names
The whole operation to build lists and append them and generate an atom
in the end is quite expensive: not only it generates a lot of lists and
copies of lists that later need to be garbage collected, but also, the
`list_to_atom` operation is a concurrent bottleneck: the atom table is
not optimised for inserts, only for reads, this one needs to grab locks
and check if the atom already existed and so on. Instead, store the name
on a persistent_term record, and simply fetch that when needed, making
the hotter part of the code, finding the pool to submit a task to,
faster.