Optimise garbage collection
Note that running garbage collection within a function will not collect
the references this function has, so the queue will actually not be
cleared after being flushed as it was expected. Instead, we use the
option `async` for the GC. This will make this process end its
reductions, and once it has been preempted, it will be scheduled for
garbage collection, and thereafter, a message will be delivered
notifying him of so.