Fix a bug where inbox passed rooms into does_user_exist
This needs to verify the type of the request at the moment where we need
to decide if the inbox owner exists. We can't return false because that
would drop the required markers from resetting the count. We can't
return `can_access_room` because then the message telling the user that
he has been removed from a room would arrive to the user at a time when
he's not a member of the room anymore. So we simply must return true
from here.