Implement a caching subsystem for muclight affiliations
It happens easily that querying the affiliations for a room becomes a
global bottleneck on a loaded server: note that every c2s process
fetches this information on their own and individually send the messages
to the given affiliated users, but when in a room of many users all send
a message at once, these affiliations are fetched as many times as users
sent a message, which is a potentially redundant overload of the DB,
more so when this operation is synchronous.

For this, we add new subscribers to the acc_room_affiliations hook: when
affiliations are to be fetched, first a cache handler is run, and if it
returns the value, the hook stops there. Otherwise, after the call to
the real DB handler, another cache handler will insert this record in
the cache, for future calls to it.

Note that we implement here other specially important hooks:
remove_domain clears up the cache on that important event, but also very
importantly forget_room and room_new_affiliations respond to the room
being removed or the affiliations changing, to keep the cache
consistent.

This has now the usual challenge: distribution, on a clustered node. The
events that keep the cache consistent, i.e., forget_room and
room_new_affiliations, are local to a single node. So either we keep
very low ttls on the cache for a fast eventual consistency, or, we
ensure the hooks are run on the whole cluster rapidly.