Do the cluster consistency
First of all, on affiliations updates, ask the cache to delete the
record from all other nodes. This will, using pg (so keep in mind the
consistency guarantees), make all other caches delete the record, to
prevent the case where other nodes will work on invalid affiliation
lists.

Then, we also add defaults, to keep the affiliations cache a live of 6
minutes. This might hopefully be a good compromise between responding to
sudden peaks in load, and reducing the risk of an inconsistent
affiliation list on some node.