Fix return value in mod_cache_users
Currently, this handler would return what the merge returns, which could
potentially break the expected value for this handler, as if the merge
fails, it would return false instead. The merge would fail in case of
some enormous lock contention, but there's retries and fast atomic
operations going on here, so chances are very very low. Still,
correctness.

Also take the chance and correct a wrongly typed spec.