Safely parse mam retractions
With the current code, a malicious retraction could have a namespace but
then not declare any id, and in that case the function would return
`{origin_id | stanza_id, none}`, which breaks the contract that the
second term must be a binary. In a worst-case scenario, this could
trigger an exception later on, maybe even killing one of the MAM
async workers with their message queues if the exception is not caught.