[wpool] send request callbacks
This allows for clients not to have to wait for the request to be
finished. It can easily happen that the workers in the pool are
overloaded, for example RDBMS workers as in this PR, so doing a
gen_server:call would block the caller until the RDBMS worker returns,
when instead, the caller can control not to make more calls until the
previous has been answered, and in the meantime, do other stuff without
being blocked.