Fix MAM cache name
When mod_mam_cache_user checks its cache, it gives its own module name
to mongoose_user_cache so that this one fetches the config to see if it
has configured its own cache or is reusing some other module's cache.
The problem is that mongoose_user_cache will check for the module key,
but it the current implementation, we are hiding it behind a cache key
to contain the cache config.

So by just moving the config a level up, mongoose_user_cache can find it.