Extend the inbox query to be able to fetch an entire inbox entry at once
This is useful for example when a conversation is entered, through some
link, to any point in the far past of the conversation, instead of at
the bottom. The inbox might be desired in a two-column layout, like in
many phone apps when displayed horizontally, but if loading from MAM at
some point far in the past, in a conversation whose inbox entry wasn't
already loaded (old conversation, pagination), building the right inbox
snippet would take many round-trips.