Add hook event on new inbox forwarded message
Custom code might want to add metadata to each inbox entry, as they
fetch it, like for example markers, or information about how to display
the entry. In order to avoid multiple round-trips between client and
server, this metadata could be added in the inbox query already. For
that, we insert a hook that takes the accumulator, the inbox row as
returned from the DB, and the IQ that initiated the query to begin with.

That way, custom xml elements can be added to the inbox iqs in the same
way they can be added to the mam queries, and hooks can parse extra
flags and generate more xml elements to append to the inbox response.