Reorder columns in the inbox schema
This doesn't make any functional changes in functioning systems, as we
never query by star but rather always name and order the queries, but
newly deployed DBs can take advantage of better aligned columns, as
padding is added to columns to align their beginning on every machine
word. This way, if a int (32bits) is followed by a varchar, the int is
padded to 64bits and only then follows the varchar, so it saves a few
bytes per entry to have for example two ints together, that way they
both fit within a single machine word and no padding is needed.