Implement the DB model for the inbox boxes
Note that prepared queries are heavily modified here. This is just to
make the ordering consistent, so it is the same in the schema
definition, in the prepared queries, in the decoding... As we don't have
a strong typing that ensures what is parsed, making similar lines not
just similar but almost identical makes it a bit easier to not miss
something out of order.

We also move the archive-to-inbox automatic logic to the database query,
as this time there's more boxes and the easiest way to know if the
preconditions apply if on the DB itself.