Declare timestamps as incremental on DB updates
In the case of mysql, such incremental field is required to be the last
one, so we need to reorder the fields here.

See https://thewebfellas.com/blog/conditional-duplicate-key-updates-with-mysql/

------------------------------------------------------------------------

> Simple conditional updates

Unfortunately there’s a catch: the app can process events out-of-order
meaning updates to the summary table won’t necessarily occur in the
order that the events are created. This means that the last_event_id and
last_event_created_at fields should only be updated if the event is
newer than the last one for the day. In SQL terms it should look like
this:

INSERT INTO daily_events
  (created_on, last_event_id, last_event_created_at)
VALUES
  ('2010-01-19', 23, '2010-01-19 10:23:11')
ON DUPLICATE KEY UPDATE
  last_event_id = VALUES(last_event_id),
  last_event_created_at = VALUES(last_event_created_at)
WHERE last_event_created_at < VALUES(last_event_created_at);

The bad news is that this won’t work as MySQL doesn’t allow WHERE
clauses in the update portion of the query. In a simple case like this
the easiest workaround is to use the IF function:

INSERT INTO daily_events
  (created_on, last_event_id, last_event_created_at)
VALUES
  ('2010-01-19', 23, '2010-01-19 10:23:11')
ON DUPLICATE KEY UPDATE
  last_event_id = IF(last_event_created_at < VALUES(last_event_created_at), VALUES(last_event_id), last_event_id),
  last_event_created_at = IF(last_event_created_at < VALUES(last_event_created_at), VALUES(last_event_created_at), last_event_created_at);

This works by checking if the last_event_created_at timestamp of the
event being updated is newer than the current timestamp, if it is then
the new value is assigned to the field in the update, otherwise the
current value is used.

An important thing to keep in mind when using this approach is that the
order in which you update your fields is very important. I was wrongly
under the impression that the updates took place in one mass-assignment
after the entire query had been interpreted by MySQL. But they’re not:
the assignments happen in the order they appear in the query. To give
you an example, this query won’t produce the expected result:

INSERT INTO daily_events
  (created_on, last_event_id, last_event_created_at)
VALUES
  ('2010-01-19', 23, '2010-01-19 10:23:11')
ON DUPLICATE KEY UPDATE
  last_event_created_at = IF(last_event_created_at < VALUES(last_event_created_at), VALUES(last_event_created_at), last_event_created_at),
  last_event_id = IF(last_event_created_at < VALUES(last_event_created_at), VALUES(last_event_id), last_event_id);

When the update is executed with a more recent event, the
last_event_created_at field will be updated, but the last_event_id field
won’t. This is because when the second IF is evaluated
last_event_created_at has already been updated so that
last_event_created_at is equal to VALUES(last_event_created_at). Crazy
huh?!

------------------------------------------------------------------------