Introduce the bin box and turn async deletes to bin moves
This fixes the fact that the asynchronous inbox is currently dangerously
inconsistent. If on two nodes one writes new messages to a user's
conversation with a room, and in another node that user is removed from
the room, the order in which those events are flushed into the DB is now
not ensured, and the risk of the removal being flushed before the
insert, and therefore the insert recovering a conversation forever, is
much higher than in the synchronous inbox.

With the concept of boxes, the async backend can instead have a deleted
box, called bin, so that removals aren't really removals, but just moves
to another box, which are atomic and can interleave with modifications
to that entry's other columns. Then asynchronously the deleted boxes can
be collected, much like email clears the Bin after so many days.