Implement HTTP endpoint to empty the bin globally and for a user
Note that the handle_result callback in the API stack had to be
modified. The HTTP standard allows to return a message body for a DELETE
call under the status code 200-OK, so we want to allow for such
possibility in the code. Two old delete operations needed to be updated
though, as they were inconsistently registering they'd return a message
that was then ignored.