Get dist metrics using 'tls_sender' to make it work for OTP 24.3.4
The previous method was very error-prone as it traversed links until
it found a port, which could be (and was in the tests) the wrong one.
Any port linked to the primary Pid would make it crash as well.

The new implementation uses the OTP internal API (undocumented),
but it is still better than following process/port links blindly.

- The dist-pid logic is used only for distribution, not always as before.
- The try-catch expressions are there primarily to catch any dying
  processes or closed ports.
  Logging has INFO level as such events are nothing serious,
  and any issue with a new OTP version would be caught by tests.