Prepare new c2s hooks and add the first handlers
Note that gen_hook type now allows to take the type for the function
handler, in order to specify stronger types. Note that dialyzer might
not be strong enough to identify this polymorphism, but it might
probably be in the future, as well as different tools like Gradualizer.

Also mongoose_acc uses erlang:system_time/1 instead of the os
equivalent, to better resist time warps and also for performance, and
likewise such time is later used to calculate a metric for the time the
message took to be handled.

Now, all c2s handling stanzas will be captured and handled in hooks, and
mongoose_acc will be a universal accumulator.