Encapsulate requests to the statem engine
The idea is that hooks now can request the gen_statem engine to do
things on return, like putting payloads on the socket to deliver to the
user, adding timeouts to the engine, adding ad-hoc events to be
triggered by the engine (see gen_statem's next_event action), or pushing
a new callback module to take control over the state machine.

This accumulator, specific to the statem engine, is embedded into the
mongoose_acc map, and then parsed and acted upon on all hook and
handle_event returns.