Simplify argument processing in mongooseimctl
The arguments were broken into pieces (on spaces),
quoted, and finally reassembled by Erlang code.

All these steps are now replaced with putting the arguments into an
array, whose elements are then passed as arguments.

One of the main causes for this change was a bug in arguments
concatenation, which was revealed by new command tests.