Handle accumulated affiliations efficiently
Affiliations where stored in the mongoose_acc for atomicity, and
performance. If a message processing pipeline in a room starts with
certain affiliations at the time of that message, we don't want to have
some handlers in the pipeline process the message with certain affs,
only to have the affiliations changed concurrently and the next handler,
for the same message, be processed with different ones.

But the code as currently written was quite redundant. The main call to
get the affiliations from the accumulator, was running the hook that
would check the acc, the cache, and the DB. We can skip that hook if
we're already in the context of muc_light and simply check if the
affiliations are already in the accumulator, and only if not then we run
the hook.

Also take the opportunity to tag the affiliations stored in the
accumulator, with the room's jid, just to avoid the (though improbable)
chance of more than one room's affiliations being accumulated under the
same key.