Integrate the domain handler with mongoose_admin_api
Error messages are changed to capitalized strings, because it is the
convention used for the whole admin API.

Error codes are left as they were except ones that are already
provided by mongoose_admin_api itself, e.g. "401 Unauthorized".

The error messages should be common for REST and GraphQL,
but this unification is left for the future.