Prepare domains backend to allow retries
Current logic for domain deletion first checks that the domain exists,
then removes the domain, and then runs the remove_domain hook. But this
is a problem, because if any of the hooks fails, then retrying the
domain removal won’t work, because it will be stopped at the first check
for domain existence.

Here we instead mark the domain as "in progress for deletion", which
will already disable it, and only finally delete it _after_ all hook
handlers have run hopefully successfully.