Return valid JSON value for Admin REST API errors
Previously it was plain text, but the Content-Type was application/json
This caused issues with the interactive Swagger UI - it displayed "no
response from server". Now it displays the errors correctly.

Tests don't need changing, because they are decoding JSON
automatically, and the result is the same text as before.