Plug stream_management into presend and create handle_flush
In the original implementation, stream_mgmt first checks on the very beginning of the route,
if there is a sid conflict. Then _all_ the `user_receive_packet` hooks are executed,
and only after that, buffering for CSI and stream_mgmt happens.

But in the current implementation we've partitioned `user_receive_packet` into all four types of
stanzas, which could potentially stop the handling, when stream_mgmt has already buffered the
payload, therefore we're having a bug.

So we add a "pre-send" event hook, where stream management will do the buffering, and a very early
stream_mgmt `user_receive_packet` will do the sid conflict check.

We also add a new kind of c2s statem event: {flush, Acc}. This will run handle_flush, which is the
second part to handle_route, meaning, after all the `user_receive_*` handlers.