Clean some details about CSI tests
This is an important part of the rework. The point is that now a session
is ready to receive messages directly addressed to it immediately after
binding a resource, so afterwards a request to enable stream management
can have interlapping messages between the request and the response. And
this is fine, as XEP-0198 requires:

- _sending_ count to start after _sending_ the <enable> or <enabled>
- _receiving_ count to start after _receiving_ the <enable> or <enabled>

So the client will set its counter for received messages after receiving
the <enabled>. So the question is then whether the order in which the
stanzas are received is relevant:

- If the client receives first the stream management <enabled> and then
  the rerouted stanzas, that means that the c2s received the <enable>
  and answered with <enabled> before receiving the rerouted stanzas, and
  therefore counted them for the sending counter.
- If the client receives first the messages from the reroute and then
  the <enabled> stanza, that means that the c2s received first all the
  messages and delivered, before receiving the <enable> and initialising
  its counters, as the client will have its received stanzas at zero
  (starting when it received the <enabled>), the server will also have
  its sent stanzas at zero (starting when it sent the <enabled>).

So in all cases the counts match.