Ensure inactive is not reordered behind messages
Many tests have Alice sending the inactive stanza, immediately followed by Bob
sending many messages. But here lies a race condition: a user message is
implemented as the test-case Pid sending Erlang messages to the user's
socket-managing Pids, which in turn send the actual TCP messages. But while the
order in which the messages sent from the test-case Pid is ensured, the order in
which the socket-managing Pids will be executed is not, and therefore it is
possible that the first message from Bob will be delivered _before_ Alice's
inactive request.

So after sending an inactive request, we need to introduce a small wait,
followed by a verification that Alice has not received any message. Still, a
wait is not bulletproof, but it solves the problem sufficiently often. The issue
is that verifying that Alice has processed the inactive request is non-trivial,
as CSI describes no answer to it.