Fix hooks for checking listener access rules
Previously the handlers were installed with the access rules
themselves, what caused:
- Possible duplication of handlers in case of same rules.
- Running all handlers sequentially, effectively performing an
intersection of all rules instead of applying rules for a given
listener.

Now the handlers are installed with the listener ID's,
and for each new session these ID's are compared with the c2s data,
resulting in only the right handler being executed.
The access rules can be just extracted from the c2s data itself.