Simplify presence handling in mod_global_distrib_SUITE
There was a function called 'wait_for_registration', that called
the non-existing old c2s module.

It was supposed to wait for the sm session, but actually the presence
is sent by the server after creating the session, not before as the
comment said.

Thus, it was enough to disable the filter predicate for Alice's
presences, and just handle them the same way as for other users.