Use mongoose_data_forms in event pusher
The form is now parsed to a map instead of a KV list, because it is
later converted to JSON, which is saved to RDBMS. The DB can alter the
order of JSON pairs when they are read, which makes pattern-matching
for already sent push notifications fail unpredictably, resulting in
double notifications, i.e. by using maps we eliminate an elusive bug.