Fix flaky sm tests by making sid's unique
Sid's should be created for different Pids - otherwise the timestamp
alone does not guarantee uniqueness.

The sm tests started failing on fast machines, which could generate a
lot of sid's in a single microsecond. On Apple M1 the small tests were
failing all the time.

To fix this, spawn a new process when generating a sid in the tests.
Forward incoming messages to the test process, because some tests need
to check them.