mod_register: Migrate from one treap in ets to all in ets
The old logic stores a single data structure in the ets table, a "treap",
which is then cleaned and scaned, but we might do with simpler code and store
the entries directly in the ets table, and clean up by full scans.

Using a single treap means that the whole data structure is copied to processes
and then garbage collection can be expensive. It also means that two processes
adding records might override each other and therefore cause inconsistencies.

The pitfall of the ets table is full scans when cleaning old timestamps, but
with low timeouts it might be fine, also considering this module should be
seldom enabled in public setups.