Unify sets in the presence state record
The idea is to keep track of a smaller set of data, I'm assuming that jids that
would fall both on pres_f and pres_t are common, so we can keep track of them
only once instead of duplicating. This way we also give a meaning to the values
of the map underlying the implementation of sets.

Note too that this set is rarely updated: it is build once upon user starting a
presence session, and it is updated only on roster changes.