Remove XEP-0018
This is functionality that was deprecated from XMPP already back in 2003 (!),
and we still have code for it that has been surviving different XEP
advertisement updates as well as shuffling of presences code.

Note that this is not related to the https://xmpp.org/extensions/xep-0126.html
XEP, as this one achieves the same by usage of privacy lists specific for the
purpose, thereby establishing only a client-based protocal that is transparent
for the server.