Make CETS node cleanup consistent
Previously, during node cleanup for CETS:
- Some modules removed only local content
- Others performed cluster-wide CETS operations

It is not possible to always remove local content because of
session_cleanup and the need for insertions - see e.g. stream management.

Meanwhile, Mnesia cleanup consistently removed all content across the
cluster.

After this change, CETS does the same.

An immediate benefit is that after the first global:trans/2 finishes
in mongoose_cleaner:run_node_cleanup/1, cluster state is consistent.

Previously cleanup would need to succeed on all nodes (sequentially),
and before that the cluster would be in an inconsistent state.