After a restart, wait until other nodes finish the cleanup
The goal is to delay node startup until other nodes perform the
cleanup. It is enough to wait for one global_trans to finish,
because the first node cleans up all the clustered CETS tables.

mongoose_cleaner is started early, so it makes sense to wait there.

Possible corner case:
- Cleanup hasn't started yet when the node is already back online.
  This is unlikely, because cleanup starts as soon as 'nodedown' is
  received.