Put CETS version in cluster name
When performing a rolling upgrade e.g. from 6.2.0 to the upcoming 6.2.1,
CETS would crash on all nodes due to version incompatibility.

This can be prevented by putting cets 'Major.Minor' version in the
cluster name, making the two cluster components disjoint during
upgrade.

The only thing to keep in mind is to bump the Major or Minor version
number when there is any incompatibility in the CETS inter-node API.