Make commit hash in system status fail-safe
This hotfix is needed to make the build for the 6.2.1 release pass.
The tests were failing only for GH tags, because the commit hash was
not available.

For now - default to an empty string (as binary, to avoid '[]' output).
We could make the field optional in the future to clean this up.