Add utilities for testing TLS error handling
The main goal is to check errors without sending the request.
Otherwise, the errors would be unpredictable due to race conditions.

For example, if there is a TLS error, and we try to send a request,
this would sometimes succeed, and sometimes fail.
Receiving the response could result in the 'closed' error or in a TLS error,
depending on the timing.

This is why neither of the clients (fusco, gun, httpc) can be used.
Instead, the 'ssl' module is used directly.