Update gen_hook tests after moving metrics to mongoose_instrument
A few tests checked that the same hook could have both
global and host-type handlers. This is not the intended behaviour, and
results in an error log.

One extra test was added for this case, while the remaining tests were
amended to test the usual case: a hook is either global or
per-host-type (not both).