Allow mod_global_distrib to be run for multiple hosts
The issue was uncovered by the small tests. Additionally, the instrumentation
is started correctly in these tests.

The host_type is a misnomer, as global_distrib does not work for host types -
this issue may be revisited. The other issue is that other events in
global_distrib are without any labels, as they were run with `global`, or can
be differentiated by the host value in the measurements map.