Test instrumentation events more correctly
Now, the tests will really fail the suite in case of failure, and they will not
pollute subsequent tests.

The contents are checked in the tests to see that the proper events are tested
and not ones from different test cases.

mod_event_pusher_rabbit has the pool tag hardcoded, and I decided not to change
that, so the messages_published_events_are_executed test can't use the pool
from the previous test.