Change event table type to duplicate_bag
The goal is to have reliable assertions for multiple events.
With the 'bag' type, an event occuring twice would be stored once
if all metric values were the same. The problem is that some metrics,
e.g. time, are unpredictable, and the resulting number of events would
be non-deterministic.

This change makes it necessary to fix multiple tests, but they are
being changed anyway - see the subsequent commits.