Fix ldap dialyzer warnings on Erlang 27
We pass anon_auth option into eldap:open/2, which is not allowed
in the dialyzer spec for this function. But this option is there
and works. Maybe OTP team forgot to add it into the spec.

So we extract options into connect_opts, which is enough to not
raise the dialyzer warning.

maybe_b2list usage in some places was adjusted to not cause warnings
in Erlang 27