Allow config processor to be a list of functions
This way, you can specify a list instead of having to merge functions

Also: remove the 'list_processor' type, because it was created when all
sections would be formatted as lists. Now they are formatted as maps
by default. What's more, if sections are merged, only the first
processor gets a map/list as the argument, and it can pass anything to
the subsequent processors.

The 'processor' type covers all such cases.