Add a test checking the case of PREPARE not followed by EXECUTE
The issue is present in CockroachDB.
It is also possible to cause it in PostgreSQL by setting the
SERIALIZABLE transaction isolation level.

This unlikely scenario was unintentionally caused by the mocked
mongoose_rbdms_backend:execute/4 in several tests.
Because worker assignment is non-deterministic, the tests were flaky.

This new test is designed to fail every time it is run,
because it enforces execution on particular workers:

1. Worker 1 calls PREPARE without EXECUTE.
2. Worker 2 calls INSERT.
3. Worker 1 calls SELECT and gets stale data,
   because it is stuck in the transaction started by PREPARE.

The test will fail (on purpose) if there is only one worker,
because it wouldn't be possible to have two distinct workers.

Details: https://www.postgresql.org/docs/current/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY