Handle race-condition of node not being validated
If it is restartating at the end of the test suite
If MIM is restarted asynchronously, simple checking if application is running
is not enough. We first need to wait for MIM to actually fully restart.

The best place to do it is in end_per_suite. So, if some suite crashes MIM,
the rest of suites would quickly fail in init_per_suite (and we need to wait
only once)