Do not persist credentials on GH Actions
I ran a GH Actions safety checker, which suggested that the
"persist-credentials" option should be set to "false".
The rationale: https://woodruffw.github.io/zizmor/audits/#artipacked. It is not
a big issue for us, as we don't upload artifacts from GH Actions, but using
this option should add a bit of safety anyway.