Fix "no such file or directory" on startup
Running this on 30e119170dd2d4b1ac44e2ea9279608826af0548 (a version of
master that passes), shows that under
`_build/mim1/rel/mongooseim/lib/mongooseim-*/priv/lib` it is expected to
find the `.so` objects of the recently removed native code files. This
is explored when the env-var removed in this commit is attempted to be
expanded. This was used until 087452273720d1f87186fda456c7e23d849c0e44
when it was removed.