Support dynamic domains for components
Note that components might register _any_ domain, even one not
necessarily provided by a static config or a dynamically configured one.
Hence it makes no sense to use the subdomains registration API as it was
described in a TODO in `mod_disco:get_external_components/2`.

Furthermore, when it comes to discovery, we want to leave all components
as discoverable, regardless of whether the component is a subdomain of a
static domain as it was previously done. The code filtering out
unrelated parent domains comes from long before MongooseIM 1.0.0, as
well as the option to set `extra_domains` to reenable visibility.