Improve the consistency of just_tls helpers
Note that now we can use `erlang:alias/1` to ensure that this message
doesn't come from anywhere else, so the `Ref` parameter is not needed to
ensure uniqueness. This allows us to unify the c2s preparing options
from all other submodules preparing so, and we can fix naming convention
by simply `make_client_opts` vs `make_server_opts`.

Note that the given `verify_fun` is just the default for the cases when
`verify_mode` is `peer` or `none`, so only for `selfsigned_peer` it is
needed to be provided explicitly. Also, the `verify_fun_wrapper` gets
quite more readable this way.