Support the calue "max" of "max_items"
Previously, the only accepted values were positive integers,
while XEP-0163 describes the field as:

    Max # of items to persist.
    `max` for no specific limit other than a server imposed maximum.

Thus, the field needs to be a string, but it requires special handling,
since anything different than "max" would be interpreted as an integer.

The "server imposed maximum" is the "max_items" value
configured in the options of mod_pubsub.