Fix collecting inet stats
The currect way of collecting metrics was failing because tls_sender:dist_tls_socket/1 is gone in OTP28, so we now grab the TCP port for TLS links by digging through sys:get_status/1. find_port/1 recursively scans the status info and returns the the underlying port when it sees it. This approach covers OTP 28 (env) and OTP 27 (static) and makes it a bit more future-proof in case there will be some changes to the status structure.