Add prefix/suffix options for sasl_external using common_name
The prefix/suffix are added to the common name in order to obtain
the user name.

The validator is 'jid_localpart', because a part of localpart needs
to be a valid localpart itself.
Early validation is easier to debug than subsequent auth failures.

Note that the section is not present by default.