Eliminate race conditions in mongoose_wpool_SUITE
The max_worker_queue_len_with_two_workers test was flaky because even
if a worker is not busy and has an empty queue, a new request occupies
the queue for a short time.

As a result, a subsequent call to mongoose_wpool:cast/4 might assume
that there are more requests in the queue than expected.

The solution is to ensure {waiting, Worker} is received before moving
forward.

The test for one worker is just updated for consistency.